feat: Complete service layer and controller implementation
- Finalized service logic and integrated corresponding controller
endpoints.
- Enhanced exception handling with structured error responses at
relevant layers.
- Developed a refined PNR generator using a dedicated utility class for
unique identification.
- Updated model classes to process `Long` values instead of `int` for
better consistency.
- Modified repository classes to align with model changes and ensure
proper data management.

+++ src/main/java/com/example/booking/trainticket/service/BookingService.java
@@ -6,7 +6,7 @@ import com.example.booking.trainticket.dto.TicketDto;
 
 public interface BookingService {
 
-	TicketDto bookTicket(int trainId, int travellerId,
+	TicketDto bookTicket(Long trainId, Long travellerId,
 			LocalDate departureDate);
 	
 	void cancelTicket(String pnr);
